requirements.txt
flask
pillow
pymongo
python-dotenv